+++ CHANGELOG.md
@@ -30,7 +30,9 @@
 - PR #72 Bug fix for segfault issue getting transpose from adjacency list
 - PR #105 Bug fix for memory leaks and python test failures
 - PR #110 Bug fix for segfault calling Louvain with only edge list
+- PR #120 Bug fix for segfault calling spectral clustering with only edge list
 - PR #116 Added netscience.mtx dataset to datasets.tar.gz
+
 
 # cuGraph 0.5.0 (28 Jan 2019)